Pin
Pin

+++ EventParkering/Model/ParkItem.cs
@@ -4,8 +4,8 @@ namespace EventParkering.Model
     public class ParkItem
     {
         public string name { get; set; }
-        public int lat { get; set; }
-        public int lon { get; set; }
+        public double lat { get; set; }
+        public double lon { get; set; }
         public string dist { get; set; }
     }
 }
